remove real connection multiplexer from unit tests, move scaler tests to integration tests
