+++ test/dotnet/Integration/IntegrationTestHelpers.cs
@@ -93,7 +93,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.Redis.Tests.Integration
         private static string GetFunctionsFileName()
         {
             string filepath = RuntimeInformation.IsOSPlatform(OSPlatform.Windows)
-                ? Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), @"npm\node_modules\azure-functions-core-tools\bin\func.exe")
+                ? @"C:\Program Files\Microsoft\Azure Functions Core Tools\func.exe"
+                //? Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), @"npm\node_modules\azure-functions-core-tools\bin\func.exe")
                 : @"/usr/bin/func"; 
             if (!File.Exists(filepath))
             {
